+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Unreleased
 
+#### Fixed
+* **\[core\]** Fixed bug that was introduced into `ArezContext.when(...)` methods that could result in an invariant
+  failure if the condition immediately resolved to true. The invariant failure was the result of
+  `Disposable.dispose(watcher)` being invoked before watcher had been assigned.
+
 ### [v0.59](https://github.com/arez/arez/tree/v0.59) (2018-02-26)
 [Full Changelog](https://github.com/arez/arez/compare/v0.58...v0.59)